감정 그래프 점유율 세팅 수정 1. 점유율이 가장 높은 감정이 2개 이상일 경우 감정 개수의 합으로 각 감정을 나눠서 점유율에 더하도록 수정 2. 점유율이 가장 높은 감정이 1개일 경우 그 감정값만 1 더하도록 수정

+++ TTogilgi/app/src/main/java/com/example/ttogilgi/model/Emotion.kt
@@ -1,9 +1,9 @@
 package com.example.ttogilgi.model
 
 data class Emotion (
-    var happiness: Int = 0,
-    var neutrality: Int = 0,
-    var sadness: Int = 0,
-    var worry: Int = 0,
-    var anger: Int = 0
+    var happiness: Double = 0.0,
+    var neutrality: Double = 0.0,
+    var sadness: Double = 0.0,
+    var worry: Double = 0.0,
+    var anger: Double = 0.0
 )
